requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
django-jazzmin==3.0.4
django-querycount==0.8.3
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
